feat: supports extends in array format
TypeScript5 starts to support multiple extends at the same time

Close #200
