+++ package.json
@@ -54,54 +54,54 @@
     "typescript"
   ],
   "dependencies": {
-    "@babel/parser": "^7.20.15",
-    "@microsoft/api-extractor": "^7.33.5",
+    "@babel/parser": "^7.21.4",
+    "@microsoft/api-extractor": "^7.34.4",
     "@rollup/pluginutils": "^5.0.2",
-    "@rushstack/node-core-library": "^3.53.2",
+    "@rushstack/node-core-library": "^3.55.2",
     "debug": "^4.3.4",
     "fast-glob": "^3.2.12",
     "fs-extra": "^10.1.0",
-    "kolorist": "^1.6.0",
+    "kolorist": "^1.7.0",
     "magic-string": "^0.29.0",
-    "ts-morph": "17.0.1"
+    "ts-morph": "18.0.0"
   },
   "devDependencies": {
-    "@babel/types": "^7.20.7",
-    "@commitlint/cli": "^17.1.2",
-    "@commitlint/config-conventional": "^17.1.0",
+    "@babel/types": "^7.21.4",
+    "@commitlint/cli": "^17.6.1",
+    "@commitlint/config-conventional": "^17.6.1",
     "@types/debug": "^4.1.7",
     "@types/fs-extra": "^9.0.13",
     "@types/minimist": "^1.2.2",
-    "@types/node": "^18.11.7",
-    "@types/prompts": "^2.4.1",
+    "@types/node": "^18.15.11",
+    "@types/prompts": "^2.4.4",
     "@types/semver": "^7.3.13",
-    "@typescript-eslint/eslint-plugin": "^5.41.0",
-    "@typescript-eslint/parser": "^5.41.0",
+    "@typescript-eslint/eslint-plugin": "^5.58.0",
+    "@typescript-eslint/parser": "^5.58.0",
     "@vue/eslint-config-standard": "^8.0.1",
     "@vue/eslint-config-typescript": "^11.0.2",
     "conventional-changelog-cli": "^2.2.2",
     "cross-env": "^7.0.3",
-    "eslint": "^8.26.0",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint": "^8.38.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-vue": "^9.6.0",
+    "eslint-plugin-vue": "^9.10.0",
     "execa": "^6.1.0",
-    "husky": "^8.0.1",
+    "husky": "^8.0.3",
     "is-ci": "^3.0.1",
-    "lint-staged": "^13.0.3",
-    "minimist": "^1.2.7",
+    "lint-staged": "^13.2.1",
+    "minimist": "^1.2.8",
     "pinst": "^3.0.0",
-    "prettier": "^2.7.1",
+    "prettier": "^2.8.7",
     "pretty-quick": "^3.1.3",
     "prompts": "^2.4.2",
     "rimraf": "^3.0.2",
-    "semver": "^7.3.8",
-    "tsx": "^3.11.0",
-    "typescript": "4.8.4",
+    "semver": "^7.4.0",
+    "tsx": "^3.12.6",
+    "typescript": "^5.0.4",
     "unbuild": "^0.9.4",
-    "vite": "^3.2.1",
-    "vitest": "^0.24.3",
+    "vite": "^3.2.5",
+    "vitest": "^0.24.5",
     "vue": "3.2.41"
   },
   "peerDependencies": {
